# api.yaml
openapi: 3.0.0
info:
  title: Delivery Management System API
  description: API для системы управления доставками
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /vehicles:
    get:
      tags:
        - Vehicles
      summary: Получить список всех машин
      responses:
        '200':
          description: Список машин
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'

    post:
      tags:
        - Vehicles
      summary: Создать новую машину
      description: Доступно только для админа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleInput'
      responses:
        '201':
          description: Машина создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '403':
          $ref: '#/components/responses/Forbidden'
  /vehicles/{id}:
    put:
      tags:
        - Vehicles
      summary: Обновить данные машины
      description: Доступно только для админа
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleInput'
      responses:
        '200':
          description: Машина обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Vehicles
      summary: Удалить машину
      description: Доступно только для админа
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Машина удалена
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Сценарий 2: Админ редактирует товары
  /products:
    get:
      tags:
        - Products
      summary: Получить список всех товаров
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      tags:
        - Products
      summary: Создать новый товар
      description: Доступно только для админа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
  /products/{id}:
    put:
      tags:
        - Products
      summary: Обновить данные товара
      description: Доступно только для админа
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Товар обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Products
      summary: Удалить товар
      description: Доступно только для админа
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Товар удален
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Сценарий 3: Админ создает и редактирует пользователей
  /users:
    get:
      tags:
        - Users
      summary: Получить список пользователей
      description: Доступно только для админа
      parameters:
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Users
      summary: Создать нового пользователя
      description: Доступно только для админа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}:
    put:
      tags:
        - Users
      summary: Обновить данные пользователя
      description: Доступно только для админа
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                login:
                  type: string
                role:
                  $ref: '#/components/schemas/UserRole'
                password:
                  type: string
                  description: Опционально, только если нужно сменить пароль
      responses:
        '200':
          description: Пользователь обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Users
      summary: Удалить пользователя
      description: Доступно только для админа
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Пользователь удален
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Сценарий 4: Курьер смотрит свои доставки
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Forbidden:
      description: Доступ запрещен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    UserRole:
      type: string
      enum:
        - admin
        - manager
        - courier
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    User:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
    UserInput:
      type: object
      required:
        - login
        - password
        - name
        - role
      properties:
        login:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        brand:
          type: string
        license_plate:
          type: string
        max_weight:
          type: number
          description: Максимальный вес в кг
        max_volume:
          type: number
          description: Максимальный объём в м³
    VehicleInput:
      type: object
      required:
        - brand
        - license_plate
        - max_weight
        - max_volume
      properties:
        brand:
          type: string
        license_plate:
          type: string
        max_weight:
          type: number
          minimum: 0
        max_volume:
          type: number
          minimum: 0
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        weight:
          type: number
          description: Вес в кг
        length:
          type: number
          description: Длина в см
        width:
          type: number
          description: Ширина в см
        height:
          type: number
          description: Высота в см
    ProductInput:
      type: object
      required:
        - name
        - weight
        - length
        - width
        - height
      properties:
        name:
          type: string
        weight:
          type: number
          minimum: 0
        length:
          type: number
          minimum: 0
        width:
          type: number
          minimum: 0
        height:
          type: number
          minimum: 0
...