# src/main/kotlin/com/example/couriermanagement/entity/User.kt
package com.example.couriermanagement.entity

import jakarta.persistence.*
import java.time.LocalDateTime

@Entity
@Table(name = "users")
data class User(
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    val id: Long = 0,

    @Column(name = "login", unique = true, nullable = false, length = 50)
    val login: String,

    @Column(name = "password_hash", nullable = false)
    val passwordHash: String,

    @Column(name = "name", nullable = false)
    val name: String,

    @Enumerated(EnumType.STRING)
    @Column(name = "role", nullable = false)
    val role: UserRole,

    @Column(name = "created_at", nullable = false)
    val createdAt: LocalDateTime = LocalDateTime.now()
)

enum class UserRole {
    admin, manager, courier
}